+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/application/CouponService.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.coupon.application;
 
-import com.woowacourse.kkogkkog.application.PushAlarmEvent;
+import com.woowacourse.kkogkkog.infrastructure.PushAlarmEvent;
 import com.woowacourse.kkogkkog.coupon.application.dto.CouponDetailResponse;
 import com.woowacourse.kkogkkog.coupon.application.dto.CouponReservationResponse;
 import com.woowacourse.kkogkkog.coupon.application.dto.CouponResponse;
